# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Orpheus Space API
  version: 0.1.0
  description: Schedule management and stored history. Catalog and result endpoints read Orpheus synchronously; writes validate new profile and template selections.
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
- url: http://localhost:8010
security:
- bearerAuth: []
- browserSession: []
- {}
...
paths:
...
  "/api/v1/auth/session":
    get:
      operationId: GetAuthSession
      summary: Get browser access state
      tags:
      - Auth
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthSession"
        default:
          "$ref": "#/components/responses/Problem"
...
  /auth/login:
    get:
      operationId: BrowserLogin
      summary: Start SP-initiated SAML login
      security: []
      description: SAML mode only; otherwise 404 auth_not_enabled. next must be a local absolute path (not auth routes); default /api/v1/auth/session. No Host or proxy header trust.
      parameters:
      - in: query
        name: next
        schema:
          type: string
          maxLength: 2048
      responses:
        "302":
          description: Start SP-initiated SAML login
          headers:
            Location:
              schema:
                type: string
              description: SAML IdP URL.
        default:
          description: Invalid return path (400), authentication disabled (404), request validation failed (422), or authentication/storage unavailable (503).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/callback:
    post:
      operationId: BrowserCallback
      summary: Consume a signed SAML HTTP-POST response
      security: []
      description: SAML mode only. Body is application/x-www-form-urlencoded with exactly one SAMLResponse and RelayState, at most 1 MiB. One-time request and browser nonce are required. Unsolicited responses are rejected.
      responses:
        "303":
          description: Consume a signed SAML HTTP-POST response
          headers:
            Location:
              schema:
                type: string
              description: Validated local return path.
        default:
          description: Invalid login state (400), invalid SAML response (401), authentication disabled (404), oversized form (413), or authentication/storage unavailable (503).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/logout:
    post:
      operationId: BrowserLogout
      summary: Revoke the local browser session
      security: []
      description: Requires the configured Origin and X-Orpheus-CSRF header equal to 1. Idempotent; clears the cookie. Does not terminate the IdP session.
      responses:
        "204":
          description: Revoke the local browser session
        default:
          description: Invalid origin or CSRF header (403), or storage unavailable (503).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /saml/metadata:
    get:
      operationId: SamlMetadata
      summary: Read service provider metadata
      security: []
      responses:
        "200":
          description: Read service provider metadata
          content:
            application/samlmetadata+xml:
              schema:
                type: string
        default:
          description: Authentication disabled (404), or metadata unavailable (503).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-orpheus_space_session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: Structured API error. 401 credentials, 403 CSRF or schedule_forbidden, 404 missing, 409
        conflict, 422 validation, 503 unavailable.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          - phase
          - details
          properties:
            code:
              type: string
            message:
              type: string
            phase:
              type: string
              nullable: true
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required:
                - path
                - code
                properties:
                  path:
                    type: array
                    items:
                      type: string
                  code:
                    type: string
...
    AuthSession:
      type: object
      additionalProperties: false
      required:
      - mode
      - authenticated
      - read_access
      - write_access
      - can_manage_all
      - user
      - expires_at
      properties:
        mode:
          type: string
          enum:
          - api_only
          - anonymous
          - saml
        authenticated:
          type: boolean
        read_access:
          type: boolean
        write_access:
          type: boolean
          description: Can create schedules and modify those available to this caller. SAML users without an email cannot write.
        can_manage_all:
          type: boolean
          description: Can manage all schedules and assign or clear their owners. True for configured SAML admins, valid Bearer keys and anonymous mode.
        user:
          type: object
          nullable: true
          additionalProperties: false
          required: [subject, display_name, email]
          properties:
            subject:
              type: string
            display_name:
              type: string
            email:
              type: string
              nullable: true
              description: Normalized email from the SAML identity, or null when unavailable.
        expires_at:
          type: string
          format: date-time
          nullable: true
...